[domain] move progress-bar into Terminal
