+++ aws-api/src/main/scala/net/kemitix/thorp/aws/api/UploadProgressLogging.scala
@@ -5,8 +5,6 @@ import net.kemitix.thorp.domain.SizeTranslation.sizeInEnglish
 import net.kemitix.thorp.domain.Terminal._
 import net.kemitix.thorp.domain.{LocalFile, Terminal}
 
-import scala.io.AnsiColor._
-
 trait UploadProgressLogging {
 
   private val oneHundredPercent = 100
@@ -23,14 +21,6 @@ trait UploadProgressLogging {
       print(s"${eraseLine}Uploading $transferred of $fileSize : $remoteKey\n$bar${cursorUp()}\r")
     } else
       print(eraseLine)
-  }
-
-  def progressBar(pos: Double, max: Double, width: Int): String = {
-    val barWidth = width - 2
-    val done = ((pos / max) * barWidth).toInt
-    val head = s"$GREEN_B$GREEN#$RESET" * done
-    val tail = " " * (barWidth - done)
-    s"[$head$tail]"
   }
 
 }
